Release/1.0.0

+++ src/Service/Feeds/Reader/FeedReader.php
@@ -50,12 +50,14 @@ class FeedReader implements FeedReaderInterface
      *
      * @throws MappingError
      */
-    public function readFeeds(int $limit = FeedReaderInterface::DEFAULT_OPTION, bool $force = false, array $feedIds = []): void
+    public function readFeeds(int $limit = FeedReaderInterface::DEFAULT_OPTION, bool $force = false, array $feedIds = []): iterable
     {
         $feeds = $this->getEnabledFeeds($limit, $force, $feedIds);
 
         foreach ($feeds as $feed) {
-            $this->readFeed($feed, $limit, $force);
+            foreach ($this->readFeed($feed, $limit, $force) as $item) {
+                yield $item;
+            }
         }
     }
 
@@ -100,14 +102,14 @@ class FeedReader implements FeedReaderInterface
                 yield;
             }
 
-            if ($feed->isSyncToFeed() && FeedReaderInterface::DEFAULT_OPTION === $limit) {
-                // We can only do cleanup if we run without limit
+            // We can only do cleanup if we run without limit
+            if (FeedReaderInterface::DEFAULT_OPTION === $limit) {
                 $this->cleanUp($feed, $start);
             }
 
             $feed->setLastRead(new \DateTimeImmutable());
             $feed->setLastReadCount($index);
-            $feed->setMessage('');
+            $feed->setMessage(null);
             $this->feedRepository->save($feed, true);
         } catch (\Exception $exception) {
             $feed->setMessage($exception->getMessage());
@@ -130,10 +132,17 @@ class FeedReader implements FeedReaderInterface
         return $index >= $limit;
     }
 
+    /**
+     * Clean up FeedItems/Events. Remove all FeedItems where
+     *  - item was not seen on last run and event is null.
+     *  - item was not seen on last run and "sync to feed is active for feed.
+     */
     private function cleanUp(Feed $feed, \DateTimeInterface $date): void
     {
         foreach ($this->feedItemRepository->findByByLastSeen($feed, $date) as $feedItem) {
-            $this->feedItemRepository->remove($feedItem);
+            if (null === $feedItem->getEvent() || $feed->isSyncToFeed()) {
+                $this->feedItemRepository->remove($feedItem);
+            }
         }
     }
 }
